pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pool-tracker"
version = "0.1.0"
description = "Read-only Hoodi staking pool tracker built on Beacon and execution RPC data."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.31",
  "typer>=0.12",
  "web3>=6.15",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools.packages.find]
include = ["pool_tracker*"]
